Use request's number of bytes to choose file size

+++ store/sqlite/sqlite.go
@@ -15,13 +15,13 @@ import (
 const (
 	timeFormat = time.RFC3339
 	// I think Chrome reads in 32768 chunks, but I haven't checked rigorously.
-	defaultChunkSize = 32768 * 10
+	defaultChunkSize = uint64(32768 * 10)
 )
 
 type (
 	Store struct {
 		ctx       *sql.DB
-		chunkSize int
+		chunkSize uint64
 	}
 
 	rowScanner interface {
@@ -35,7 +35,7 @@ func New(path string, optimizeForLitestream bool) Store {
 
 // NewWithChunkSize creates a SQLite-based datastore with the user-specified
 // chunk size for writing files. Most callers should just use New().
-func NewWithChunkSize(path string, chunkSize int, optimizeForLitestream bool) Store {
+func NewWithChunkSize(path string, chunkSize uint64, optimizeForLitestream bool) Store {
 	log.Printf("reading DB from %s", path)
 	ctx, err := driver.Open(path, blobio.Register)
 	if err != nil {
